pgf-server HTTP mode: omit charset from ContentType for static files
Ignore-this: 3468fa179ebb976b1d4a53aca10e04a7
When pfg-server servers a static file in HTTP mode, it is output as is, so
specifying charset="iso-8859-1" could be wrong.

Note: the a charset can be specified with a meta tag in HTML files,
e.g. for files in UTF-8:

<meta http-equiv="Content-type" content="text/html;charset=UTF-8">

darcs-hash:20100924123438-bae10-61d14f1d95883f23027e4664ba395845858d6450.gz