minibar.html: specify charset, move initialization
Ignore-this: 8f92a33b2725775fc26dd117edee5ca0
Added a meta tag to specify that the file is in UTF-8. This makes the browser
(at least Firefox) assume that the JavaScript files are also in UTF-8, it seems.

Moved the call to start_minibar() from the <body onload="..."> to the end of
the file. The onload action is not executed until all images in the document
has been loaded, while javascipt at the end of the file is executed as soon
as the HTML file has been loaded. (Calling start_minibar() from the head of
the document would be too soon, since it refers to elements in the body,
which hasn't been loaded yet.)

darcs-hash:20100924124718-bae10-c633e109d5d8a96a453d74d4e556f5e4ecc05d65.gz