gf-server.cabal: Change the default for the flag fastcgi to False
Ignore-this: 2216f17435b944f74699b058dac93396
If you want the fastcgi server, use cabal install -ffastcgi.

This is to avoid problems when you only want to compile pgf-http and the
fastcgi library fails to install, which it usually does, since it depends on
a C library which is not installed automatically with cabal.

darcs-hash:20101222144917-bae10-67493f2707eea36e708be245a74734b5bf8b40ba.gz