pgf-http: fix a bug that caused "+" to be treaded as " " in PGF service requests URLs.
Ignore-this: 34fbc7c854104ed4047cc1926d1dd28e
This was a bug in my workaround for a bug in the httpd-shed package. It
made it impossible to use the glue token "&+" for Turkish input in the minibar,
for example.

darcs-hash:20110825165817-bae10-22bbb9cf2f2d792336a57547ad67ebe89b003dad.gz