minibar: word-for-word replacements: use concrete syntax for replacement words when possible
Ignore-this: 21f29c3d8b5e6c345eb829198d04f499
Instead of showing the name of a function in the abstract syntax, linearize it
and show the result. For functions with argument, e.g. That : Kind -> Item,
the function is applied to the right number of placeholder arguments: 'That ?'.
If the linearization fails, the name of the function is shown anyway.

darcs-hash:20120427140001-bae10-61c96689162a8aea21f53a4de26e0816ef0f3a61.gz