GF.Command.Commands: allCommands is now a constant
Ignore-this: 64a2638b01e6862d5c777667ca83155e
The dependency on PGFEnv has been moved from the list to the exec function of
the commands in the list. This means that the help command no longer needs
to generate a new list of commands and that the state of the shell
(type GF.Command.Interpreter.CommandEnv) no longer needs to contain the list
of commands.

darcs-hash:20120925114232-bae10-6413c778f49e1f2b1de12f1b08450673541b1d23.gz