GF shell, show_operations: also show the types of linearization functions
Ignore-this: 417b5a073b16d4d47b8f4d17cfbd01ab
This is a simple change in GF.Grammar.Lookup.allOpers, which is used only in
the implementation of the show_operations command in the shell.

This is useful when importing a concrete syntax (like LexiconEng) as a resource.
However, the types don't always look as nice as I hoped...

darcs-hash:20120927155520-bae10-4ccbe534cd4d789b4e663195d1bf082f92020d91.gz