GF.Command.Command: turn CommandOutput into a newtype
Ignore-this: 1dbb6fd3b37f3a5850b93b8046fc6c99
The output from commands is represented as ([Expr],String), where the [Expr] is
used when data is piped between commands and the String is used for the final
output. The String can represent the same list of trees as the [Expr] and/or
contain diagnostic information.

Sometimes the data that is piped between commands is not a list of trees, but
e.g. a string or a list of strings. In those cases, functions like fromStrings
and toStrings are used to encode the data as a [Expr].

This patch introduces a newtype for CommandOutput and collects the functions
dealing with command output in one place to make it clearer what is going on.
It also makes it easier to change to a more direct representation of piped
data, and make pipes more "type safe", if desired.

darcs-hash:20121016130103-bae10-9da84d39437de6733a65d5bed31551f8b9422866.gz