Syntax editor: in-place replacement of functions
Ignore-this: 5c9c6bec66e73ddaf9dbbe023e2829a7
When at a non-leaf node, refinements with identical type signatures
are highlighting and can re placed without destroying the children.
If not, the refinement is greyed and the user is asked to clear
the current subtree first if they wish to replace it.
This aspect of the UI should be polished, but at least it is obvious.
Also, some substantial optimizations can still be made to cache
the processed type signatures (which determine what can be replaced
in-place)

darcs-hash:20121130105642-70c34-79c1e3ea4343753ed59e56ecaf475a8599f0232e.gz